Allow IPv6 resolver addr, instead of 127.0.0.11.
May be useful in an IPv6 only network (although, maybe not, does Linux
always have a "127.0.0.0/8"?).

Any LL address can be used, but any may clash with an IPAM pool or the
standard LL prefix. That probably doesn't matter, even if it happened to
be an address on "lo", the problem would only come with something
listening on port 53.

Tests not updated, a couple of TODO(robmry)'s left in.