libn/osl: work around kernel bug
When the net.ipv4.conf.[iface].arp_notify sysctl is set to 1, the
kernel is supposed to send a gratuitous ARP when the interface is
brought up. However, this functionality is broken on at least the
Ubuntu 20.04 kernel
  https://bugs.launchpad.net/ubuntu/+source/linux/+bug/1989187
and Docker Desktop's 6.6.16-linuxkit kernel.

So, trigger the kernel to send a gratuitous ARP using the other
event: setting the interface's MAC address after the kernel has
finished bringing up the link asynchronously.

Co-authored-by: Cory Snider <csnider@mirantis.com>
Signed-off-by: Rob Murray <rob.murray@docker.com>