Configurable count and interval for gratuitous ARP/NA messages
The default is to send 3 messages at 1s intervals.

That can be overridden in "docker network create" using:
  -o com.docker.network.adv_addr_count=3
  -o com.docker.network.adv_addr_ms=1000

Or, in daemon.json for each driver:
  "default-network-opts": {
    "bridge": {
      "com.docker.network.adv_addr_count": "3",
      "com.docker.network.adv_addr_ms": "1000"
    }
  }

The allowed range is 0-3 for count, and 100ms-2s for the interval.
Setting count to 0 disables the gratuitous ARP/NA messages.

The default bridge will always use the built-in defaults, it cannot
be configured.

Signed-off-by: Rob Murray <rob.murray@docker.com>